requirements.txt
# ==================== FastAPI 核心 ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0

# ==================== 安全与限流 ====================
slowapi==0.1.9

# ==================== 工具 ====================
python-dotenv==1.0.1
python-multipart==0.0.6
psutil==5.9.8

# ==================== RAG 核心依赖 ====================
# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# LLM API Clients
openai==1.10.0
anthropic==0.18.0
zhipuai==2.0.1
dashscope==1.14.0

# ML Dependencies
numpy>=1.24.0,<2.0.0
torch>=2.0.0
transformers==4.36.2
scikit-learn==1.3.2

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2

# Utilities
tqdm==4.66.1
requests==2.31.0

# BM25 for Hybrid Retrieval
rank-bm25>=0.2.1
jieba>=0.42.1
